requirements.txt
Flask==3.0.2
Flask-RESTful==0.3.10
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.6
python-dotenv==1.0.1
pytz==2024.1
referencing==0.34.0
rpds-py==0.18.0
six==1.16.0
SQLAlchemy==2.0.29
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
Werkzeug==3.0.1